PostLike,Hashtag Refactor
