+++ KnockKnock/src/main/java/com/example/knockknock/domain/hashtag/repository/HashtagRepository.java
@@ -1,6 +1,6 @@
-package com.example.knockknock.domain.post.repository;
+package com.example.knockknock.domain.hashtag.repository;
 
-import com.example.knockknock.domain.post.entity.Hashtag;
+import com.example.knockknock.domain.hashtag.entity.Hashtag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
