+++ KnockKnock/src/main/java/com/example/knockknock/domain/post/controller/PostController.java
@@ -1,18 +1,15 @@
 package com.example.knockknock.domain.post.controller;
 
+import com.example.knockknock.domain.hashtag.dto.HashtagRegisterRequestDto;
 import com.example.knockknock.domain.post.dto.request.*;
-import com.example.knockknock.domain.post.dto.response.PostDetailResponseDto;
 import com.example.knockknock.domain.post.service.PostService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @Slf4j
 @RequestMapping("/post")
@@ -55,40 +52,7 @@ public class PostController {
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @PostMapping("/{userId}/like/{postId}")
-    public ResponseEntity likePost(
-            @PathVariable Long userId,
-            @PathVariable Long postId
-    ) {
-        postService.likePost(userId, postId);
-        return ResponseEntity.ok().build();
-    }
 
-    @DeleteMapping("/{userId}/dislike/{postId}")
-    public ResponseEntity deletePostLike(
-            @PathVariable Long userId,
-            @PathVariable Long postId
-    ) {
-        postService.deletePostLike(userId, postId);
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/hashtag/{postId}")
-    public ResponseEntity addHashtag(
-            @PathVariable Long postId,
-            @RequestBody HashtagRegisterRequestDto request
-    ) {
-        postService.addHashtag(postId, request);
-        return new ResponseEntity<>(HttpStatus.CREATED);
-    }
-
-    @DeleteMapping("/hashtag/delete/{postId}")
-    public ResponseEntity deleteHashtag(
-            @PathVariable Long postId
-    ) {
-        postService.deleteHashtag(postId);
-        return ResponseEntity.ok().build();
-    }
 
     @GetMapping("/share/{postId}")
     public ResponseEntity<String> sharePost(
