+++ KnockKnock/src/main/java/com/example/knockknock/domain/post/entity/Post.java
@@ -2,8 +2,10 @@ package com.example.knockknock.domain.post.entity;
 
 import com.example.knockknock.domain.comment.entity.Comment;
 import com.example.knockknock.domain.board.entity.BoardType;
+import com.example.knockknock.domain.hashtag.entity.Hashtag;
 import com.example.knockknock.domain.member.entity.Member;
 import com.example.knockknock.domain.post.dto.request.PostUpdateRequestDto;
+import com.example.knockknock.domain.postlike.entity.PostLike;
 import com.example.knockknock.domain.report.entity.Report;
 import com.example.knockknock.global.timestamp.TimeStamped;
 import jakarta.persistence.*;
