+++ KnockKnock/src/main/java/com/example/knockknock/domain/postlike/repository/PostLikeRepository.java
@@ -1,8 +1,8 @@
-package com.example.knockknock.domain.post.repository;
+package com.example.knockknock.domain.postlike.repository;
 
 import com.example.knockknock.domain.post.entity.Post;
 import com.example.knockknock.domain.member.entity.Member;
-import com.example.knockknock.domain.post.entity.PostLike;
+import com.example.knockknock.domain.postlike.entity.PostLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
